+++ saleor/payment/gateways/dummy/tests/test_dummy.py
@@ -2,6 +2,7 @@ from decimal import Decimal
 
 import pytest
 
+from .....plugins.manager import get_plugins_manager
 from .... import ChargeStatus, PaymentError, TransactionKind, gateway
 
 
@@ -12,7 +13,9 @@ def setup_dummy_gateway(settings):
 
 
 def test_authorize_success(payment_dummy):
-    txn = gateway.authorize(payment=payment_dummy, token="Fake")
+    txn = gateway.authorize(
+        payment=payment_dummy, token="Fake", manager=get_plugins_manager()
+    )
     assert txn.is_success
     assert txn.kind == TransactionKind.AUTH
     assert txn.payment == payment_dummy
@@ -40,14 +43,18 @@ def test_authorize_failed(is_active, charge_status, payment_dummy):
     payment.charge_status = charge_status
     payment.save()
     with pytest.raises(PaymentError):
-        txn = gateway.authorize(payment=payment, token="Fake")
+        txn = gateway.authorize(
+            payment=payment, token="Fake", manager=get_plugins_manager()
+        )
         assert txn is None
 
 
 def test_authorize_gateway_error(payment_dummy, monkeypatch):
     monkeypatch.setattr("saleor.payment.gateways.dummy.dummy_success", lambda: False)
     with pytest.raises(PaymentError):
-        txn = gateway.authorize(payment=payment_dummy, token="Fake")
+        txn = gateway.authorize(
+            payment=payment_dummy, token="Fake", manager=get_plugins_manager()
+        )
         assert txn.kind == TransactionKind.AUTH
         assert not txn.is_success
         assert txn.payment == payment_dummy
@@ -56,7 +63,7 @@ def test_authorize_gateway_error(payment_dummy, monkeypatch):
 def test_void_success(payment_txn_preauth):
     assert payment_txn_preauth.is_active
     assert payment_txn_preauth.charge_status == ChargeStatus.NOT_CHARGED
-    txn = gateway.void(payment=payment_txn_preauth)
+    txn = gateway.void(payment=payment_txn_preauth, manager=get_plugins_manager())
     assert txn.is_success
     assert txn.kind == TransactionKind.VOID
     assert txn.payment == payment_txn_preauth
@@ -85,14 +92,14 @@ def test_void_failed(is_active, charge_status, payment_dummy):
     payment.charge_status = charge_status
     payment.save()
     with pytest.raises(PaymentError):
-        txn = gateway.void(payment=payment)
+        txn = gateway.void(payment=payment, manager=get_plugins_manager())
         assert txn is None
 
 
 def test_void_gateway_error(payment_txn_preauth, monkeypatch):
     monkeypatch.setattr("saleor.payment.gateways.dummy.dummy_success", lambda: False)
     with pytest.raises(PaymentError):
-        txn = gateway.void(payment=payment_txn_preauth)
+        txn = gateway.void(payment=payment_txn_preauth, manager=get_plugins_manager())
         assert txn.kind == TransactionKind.VOID
         assert not txn.is_success
         assert txn.payment == payment_txn_preauth
@@ -103,7 +110,11 @@ def test_void_gateway_error(payment_txn_preauth, monkeypatch):
     [("98.40", ChargeStatus.FULLY_CHARGED), (70, ChargeStatus.PARTIALLY_CHARGED)],
 )
 def test_capture_success(amount, charge_status, payment_txn_preauth):
-    txn = gateway.capture(payment=payment_txn_preauth, amount=Decimal(amount))
+    txn = gateway.capture(
+        payment=payment_txn_preauth,
+        manager=get_plugins_manager(),
+        amount=Decimal(amount),
+    )
     assert txn.is_success
     assert txn.payment == payment_txn_preauth
     payment_txn_preauth.refresh_from_db()
@@ -130,14 +141,18 @@ def test_capture_failed(
     payment.charge_status = charge_status
     payment.save()
     with pytest.raises(PaymentError):
-        txn = gateway.capture(payment=payment, amount=amount)
+        txn = gateway.capture(
+            payment=payment, manager=get_plugins_manager(), amount=amount
+        )
         assert txn is None
 
 
 def test_capture_gateway_error(payment_txn_preauth, monkeypatch):
     monkeypatch.setattr("saleor.payment.gateways.dummy.dummy_success", lambda: False)
     with pytest.raises(PaymentError):
-        txn = gateway.capture(payment=payment_txn_preauth, amount=80)
+        txn = gateway.capture(
+            payment=payment_txn_preauth, manager=get_plugins_manager(), amount=80
+        )
         assert txn.kind == TransactionKind.CAPTURE
         assert not txn.is_success
         assert txn.payment == payment_txn_preauth
@@ -165,7 +180,11 @@ def test_refund_success(
     payment.charge_status = ChargeStatus.FULLY_CHARGED
     payment.captured_amount = initial_captured_amount
     payment.save()
-    txn = gateway.refund(payment=payment, amount=Decimal(refund_amount))
+    txn = gateway.refund(
+        payment=payment,
+        manager=get_plugins_manager(),
+        amount=Decimal(refund_amount),
+    )
 
     payment.refresh_from_db()
     assert txn.kind == TransactionKind.REFUND
@@ -194,7 +213,11 @@ def test_refund_failed(
     payment.captured_amount = Decimal(initial_captured_amount)
     payment.save()
     with pytest.raises(PaymentError):
-        txn = gateway.refund(payment=payment, amount=Decimal(refund_amount))
+        txn = gateway.refund(
+            payment=payment,
+            manager=get_plugins_manager(),
+            amount=Decimal(refund_amount),
+        )
         assert txn is None
 
 
@@ -205,7 +228,11 @@ def test_refund_gateway_error(payment_txn_captured, monkeypatch):
     payment.captured_amount = Decimal("80.00")
     payment.save()
     with pytest.raises(PaymentError):
-        gateway.refund(payment=payment, amount=Decimal("80.00"))
+        gateway.refund(
+            payment=payment,
+            manager=get_plugins_manager(),
+            amount=Decimal("80.00"),
+        )
 
     payment.refresh_from_db()
     txn = payment.transactions.last()