+++ saleor/plugins/manager.py
@@ -157,18 +157,10 @@ class PluginsManager(PaymentInterface):
             )
             for line_info in lines
         ]
-        default_value = base_calculations.base_checkout_subtotal(
-            line_totals, checkout_info.checkout.currency
-        )
+        currency = checkout_info.checkout.currency
+        total = sum(line_totals, zero_taxed_money(currency))
         return quantize_price(
-            self.__run_method_on_plugins(
-                "calculate_checkout_subtotal",
-                default_value,
-                checkout_info,
-                lines,
-                address,
-                discounts,
-            ),
+            total,
             checkout_info.checkout.currency,
         )
 
@@ -593,15 +585,20 @@ class PluginsManager(PaymentInterface):
         }
 
     def list_payment_gateways(
-        self, currency: Optional[str] = None, active_only: bool = True
+        self,
+        currency: Optional[str] = None,
+        checkout: Optional["Checkout"] = None,
+        active_only: bool = True,
     ) -> List["PaymentGateway"]:
         payment_plugins = self.list_payment_plugin(active_only=active_only)
         # if currency is given return only gateways which support given currency
         gateways = []
         for plugin in payment_plugins.values():
-            gateway = plugin.get_payment_gateway(currency=currency, previous_value=None)
-            if gateway:
-                gateways.append(gateway)
+            gateways.extend(
+                plugin.get_payment_gateways(
+                    currency=currency, checkout=checkout, previous_value=None
+                )
+            )
         return gateways
 
     def list_external_authentications(self, active_only: bool = True) -> List[dict]:
@@ -614,20 +611,6 @@ class PluginsManager(PaymentInterface):
             for plugin in plugins
             if auth_basic_method in type(plugin).__dict__
         ]
-
-    def checkout_available_payment_gateways(
-        self,
-        checkout: "Checkout",
-    ) -> List["PaymentGateway"]:
-        payment_plugins = self.list_payment_plugin(active_only=True)
-        gateways = []
-        for plugin in payment_plugins.values():
-            gateway = plugin.get_payment_gateway_for_checkout(
-                checkout, previous_value=None
-            )
-            if gateway:
-                gateways.append(gateway)
-        return gateways
 
     def __run_payment_method(
         self,